Falta probar el endpoint

+++ src/main/kotlin/ar/edu/unq/desapp/grupoA/CryptoExchangeApi/model/Active.kt
@@ -2,7 +2,7 @@ package ar.edu.unq.desapp.grupoA.CryptoExchangeApi.model
 
 class Active(
     val crypto: String,
-    val criptoNominalQuantity: Int,
-    val usdPrice: Int,
-    val argPrice: Int,
+    val criptoNominalQuantity: Float,
+    val usdPrice: Float,
+    val argPrice: Float,
 ) {}
