+++ src/main/kotlin/ar/edu/unq/desapp/grupoA/CryptoExchangeApi/model/Intention.kt
@@ -14,7 +14,7 @@ import java.time.LocalDateTime
 class Intention(
     @ManyToOne(fetch = FetchType.EAGER)
     val crypto: Crypto,
-    val criptoNominalQuantity: Int,
+    val criptoNominalQuantity: Float,
     val intentionCryptoPrice : Float,
     val operation: IntentionType,
     @ManyToOne(fetch = FetchType.EAGER)
